ensure embeddings /export route is protected
